requirements.txt
# --- core bot ---
Flask==3.1.0
gunicorn==23.0.0
slack-bolt==1.23.0
slack_sdk==3.35.0
requests==2.32.3
python-dotenv==1.1.0          # if you load .env locally

# --- Gmail / email support (keep these if email_sender.py still uses Gmail API) ---
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-api-python-client==2.70.0
httplib2==0.22.0

# --- data utilities you **actually use** ---
pandas==2.2.3                 # keep only if your code imports pandas
numpy==2.2.3                  # same
openpyxl==3.1.5               # same

# --- anything else you KNOW your remaining code imports ---
